fix: replace Attrs with args
